Moved System.Drawing dependent methods to extension class. Issue #83
